pyproject.toml
[tool.poetry]
name = "transmission-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Vadim Aleksandrov <valeksandrov@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
emoji = "^2.8.0"
python-telegram-bot = {extras = ["socks", "job-queue"], version = "^20.4"}
PyYAML = "^6.0.1"
transmissionrpc = "^0.11"
aiosqlite = "^0.19.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.285"
pyright = "^1.1.324"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "E", "W", "C90",
          "I", "N", "S", "B", "A",
          "ISC", "T20", "Q", "PTH"]

ignore = ["A003"]
line-length = 120

[tool.black]
target-version = ['py311']
line-length = 120

[tool.pyright]
reportUnnecessaryTypeIgnoreComment="warning"
